finished 30 delete clean up to run on launch and RecycleBinView appear

+++ Date Tracker/Core/Core Data/Item+CoreDataProperties.swift
@@ -23,7 +23,7 @@ extension Item {
     @NSManaged public var name: String?
     @NSManaged public var taggedForDelete: Bool
     @NSManaged public var timestamp: Date?
-    @NSManaged public var timeUntilHardDelete: Int64
+    @NSManaged public var timeUntilHardDelete: Double
 
 }
